% Generated manually for package scaffold.
\name{hello_material_legacy}
\alias{hello_material_legacy}
\title{Say hello to the material legacy project}
\usage{
hello_material_legacy(name = "LTER")
}
\arguments{
\item{name}{A character string with a name to greet.}
}
\value{
A character string with a friendly greeting.
}
\description{
Say hello to the material legacy project.
}
\examples{
hello_material_legacy("LTER")
}
