% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macabre-package.R
\docType{package}
\name{macabre-package}
\alias{macabre}
\alias{macabre-package}
\title{macabre: Ecological legacies of dead foundation species}
\description{
macabre supports harmonization and analysis workflows for a synthesis project
examining how the material remains of foundation species influence ecological
processes across ecosystems.
}
\keyword{internal}
