% Generated by hand for documentation-only update
\name{ml_new_site_file}
\alias{ml_new_site_file}
\title{Create a new site standardizer file}
\description{
Creates a new site standardizer file from the template in \code{R/sites/}.
This is for site contributors. The function does not register the site.
}
\usage{
ml_new_site_file(site_id, path = "R/sites", open = FALSE)
}
\arguments{
  \item{site_id}{Site identifier (uppercase letters, numbers, underscores).}
  \item{path}{Directory where the new site file should be created.}
  \item{open}{Logical; if TRUE, prints the file path.}
}
\value{
The path to the newly created file.
}
